linux/efa: Prevent overflow on sge total bytes calculation
SGE total bytes calculation accumulated the SGE lengths into a u32. When
the sum of the lengths exceeded U32_MAX the accumulator wrapped, so the
returned total could be small enough to pass the inline data length
check.

Accumulate into a u64 instead. num_sge is an int and each length is a
u32, so the sum is bounded by (2^31 - 1) * (2^32 - 1), which is half of
U64_MAX, and can no longer wrap.

Fixes: 30efa4f8f24e ("efa: Add kernel verbs support")

+++ kernel/linux/efa/src/efa_data_verbs.c
@@ -67,9 +67,9 @@ static inline struct efa_ah *to_eah(struct ib_ah *ibah)
 	return container_of(ibah, struct efa_ah, ibah);
 }
 
-static u32 efa_sge_total_bytes(const struct ib_sge *sg_list, int num_sge)
+static u64 efa_sge_total_bytes(const struct ib_sge *sg_list, int num_sge)
 {
-	u32 bytes = 0;
+	u64 bytes = 0;
 	int i;
 
 	for (i = 0; i < num_sge; i++)
